More moving around of code; e.g., DTreeVizRender goes to utils.py and I moved the DTreeVizAdaptor to the top of the file
