+++ dtreeviz/models/shadow_decision_tree.py
@@ -1,5 +1,4 @@
 from abc import ABC, abstractmethod
-from collections.abc import Sequence
 from numbers import Number
 from typing import List, Tuple, Mapping
